Move network_test_class and visuals to modules. Refactor network_cli and run_test.
